app/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core>=0.2.38,<0.3.0
langchain-text-splitters==0.2.2
langchain-google-genai==1.0.8
langchainhub==0.1.20
langgraph==0.2.23

python-dotenv==1.0.1
fastapi==0.111.1
sse_starlette==2.1.2
protobuf>=3.19.5,<5.0.0
tqdm==4.66.4
tiktoken==0.7.0
tavily-python==0.3.4
pytest==8.2.2
python-multipart>=0.0.7

qdrant-client>=1.12.0
langchain-qdrant>=0.1.0
pypdf==4.0.1

# PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
email-validator==2.1.0
